+++ src/main/java/fuck/manthe/nmsl/entity/dto/CrackedUserDTO.java
@@ -11,5 +11,5 @@ public class CrackedUserDTO {
     private Long expire;
 
     // analysis
-    private Long totalLaunch;
+    private int totalLaunch;
 }
